// src/designer/operate-provider/right-click-menu/ContextMenu.type.ts
export enum ContextMenuType {
    /**
     * 锁定
     */
    DO_LOCK = 'doLock',
    /**
     * 解锁
     */
    DO_UNLOCK = 'doUnLock',
    /**
     * 隐藏
     */
    DO_HIDE = 'doHide',
    /**
     * 复制
     */
    DO_COPY = 'doCopy',
    /**
     * 置顶
     */
    TO_TOP = 'toTop',
    /**
     * 置底
     */
    TO_BOTTOM = 'toBottom',
    /**
     * 删除
     */
    DO_DELETE = 'doDelete',
    /**
     * 编组
     */
    DO_GROUPING = 'doGrouping',
    /**
     * 取消编组
     */
    DO_UNGROUPING = 'doUnGrouping',
}